backend/pyproject.toml
[project]
name = "snowloader-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Tom Gebicke",email = "tom.gebicke@datacoders.io"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.121.1,<0.122.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "cryptography (>=46.0.3,<47.0.0)",
    "boto3 (>=1.40.71,<2.0.0)",
    "snowflake-connector-python (>=4.0.0,<5.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pyyaml (>=6.0.1,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
